+++ packages/eth-common/webpack.config.js
@@ -102,7 +102,7 @@ function getConfig(isProduction) {
             },
             modules: [
                 sourceRoot,
-                nodeModulesPath
+                "node_modules"
             ],
             extensions: ['.webpack.js', '.web.js', '.ts', '.tsx', '.js']
         },