+++ src/app/eth-common/module/toolbar/alethioApiModule.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { IModuleDef } from "plugin-api/IModuleDef";
 import { ToolbarItem } from "@alethio/ui/lib/layout/toolbar/ToolbarItem";
 import { ITranslation } from "plugin-api/ITranslation";
-import { ExternalLink } from "plugin-api/component/ExternalLink";
+import { Link } from "plugin-api/component/Link";
 import { ProductIconButton } from "@alethio/ui/lib/layout/toolbar/ProductIconButton";
 import { AlethioApiIcon } from "@alethio/ui/lib/icon/AlethioApiIcon";
 
@@ -15,9 +15,9 @@ export const alethioApiModule: (alethioApiUrl: string) => IModuleDef<IAlethioApi
     dataAdapters: [],
     getContentComponent: async () => (props) => (
         <ToolbarItem title={props.translation.get("toolbar.api.label")}>
-            <ExternalLink href={alethioApiUrl} rel="noopener noreferrer">
+            <Link to={alethioApiUrl}>
                 <ProductIconButton Icon={AlethioApiIcon} />
-            </ExternalLink>
+            </Link>
         </ToolbarItem>
     ),
     getContentProps : data => data,