+++ src/app/eth-common/module/topbar/alethioReportsModule.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { IModuleDef } from "plugin-api/IModuleDef";
-import { MenuItem } from "plugin-api/component/topbar/MenuItem";
+import { MobileMenuItem } from "@alethio/ui/lib/layout/topbar/MobileMenuItem";
 import { ITranslation } from "plugin-api/ITranslation";
-import { ExternalLink } from "plugin-api/component/ExternalLink";
+import { Link } from "plugin-api/component/Link";
 import { ToolbarIconButton } from "@alethio/ui/lib/layout/toolbar/ToolbarIconButton";
 import { AlethioReportsIcon } from "@alethio/ui/lib/icon/AlethioReportsIcon";
 
@@ -15,11 +15,11 @@ export const alethioReportsModule: (alethioReportsUrl: string) => IModuleDef<IAl
     contextType: {},
     dataAdapters: [],
     getContentComponent: async () => (props) => (
-        <MenuItem title={props.translation.get("toolbar.reports.label")}>
-            <ExternalLink href={alethioReportsUrl} rel="noopener noreferrer">
+        <MobileMenuItem title={props.translation.get("toolbar.reports.label")}>
+            <Link to={alethioReportsUrl}>
                 <ToolbarIconButton Icon={AlethioReportsIcon} />
-            </ExternalLink>
-        </MenuItem>
+            </Link>
+        </MobileMenuItem>
     ),
     getContentProps : data => data
 });