+++ packages/eth-memento/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- **Breaking change** Update to plugin-api@1.0.0-beta.7 (@alethio/cms@1.0.0-beta.10)
+
 # v2.0.1
 
 - Fix errors caused by some optional adapters being undefined